Visualize T cell clonal expansion with a ball-packing plot.

Integrates a cell ranger T cell library into a Seurat object with a UMAP reduction. Then gets sizes of unique clones and utilizes a circle-packing algorithm to pack circles representing individual clones in approximately the same UMAP coordinates and clusters into a ggplot object.

Usage

scballpack(
  seurat_obj,
  tcr_df,
  rad_decrease = 0,
  res = 360,
  ORDER = TRUE,
  try_place = TRUE,
  progbar = TRUE,
  repulse = FALSE,
  repulsion_threshold = 1,
  repulsion_strength = 0.05,
  max_repulsion_iter = 100,
  use_default_theme = TRUE,
  show_origin = FALSE,
  clone_scale_factor = 0.01,
  retain_axis_scales = TRUE
)

Arguments

seurat_obj

Seurat object with a UMAP reduction.

tcr_df

Dataframe of the T cell library generated by Cell Ranger

rad_decrease

numeric. indicates how much the radii of the clones should decrease to add a slight gap between all of them. Defaults to 0 but 0.05 - 0.1 values are recommended.

res

The number of points on the generated path per full circle.

ORDER

logical. Decides if the largest clones should be at the cluster centroids

try_place

If `TRUE`, always minimizes distance from a newly placed circle to the origin

progbar

logical. Decides if visual cues print to the REPL of the packing progress

repulse

If `TRUE`, will attempt to push overlapping clusters away from each other. CUrrently is bugged

repulsion_threshold

numeric. The radius that cluster overlap is acceptable

repulsion_strength

numeric. The smaller the value the less the clusters repulse each other

max_repulsion_iter

numeric. The number of repulsion iterations.

use_default_theme

If `TRUE`, the resulting plot will have the same theme as the seurat UMAP

show_origin

logical. If `TRUE`, only the centers of each circle will be plotted

clone_scale_factor

numeric. Decides how much to scale each circle. Usually should be kept at around 0.01 to somewhat maintain UMAP structure

retain_axis_scales

If `TRUE`, approximately maintains the axis scales of the original UMAP

Value

Returns a ggplot2 object of the ball packing plot. Can be operated on like normal ggplot objects

See also

Examples

if (FALSE) {
library(Seurat)
library(readr)
library(scballpack)

# run sc-RNAseq pipeline to produce a seurat_object

seurat_object <- RunUMAP(seurat_object)
TCR_dataframe <- read.csv("file_location/all_contig_annotations.csv")

# produce and show the ball-packing plot
ball_pack_plot <- scballpack(seurat_object, TCR_dataframe)
ball_pack_plot
}